+++ resources/assets/js/views/related_event/index.vue
@@ -15,7 +15,7 @@
                     <router-link class="edit-modal btn btn-success" :to="'/related_event/' + item.id + '/edit'">
                         <span class="glyphicon glyphicon-edit"></span> {{localization.edit}}
                     </router-link>
-                    <button class="edit-modal btn btn-danger" @click="deleteItem(item.id,index);props.model.data.splice(index,1);props.model.total <= 10 ? showfooter = false : showfooter = true;">
+                    <button class="edit-modal btn btn-danger" @click="deleteItem(item.id,index, props.model.data); props.model.total <= 10 ? showfooter = false : showfooter = true">
                         <span class="glyphicon glyphicon-trash"></span> {{localization.delete}}
                     </button>
                 </td>
@@ -69,12 +69,14 @@
             DataViewer
         },
         methods: {
-            deleteItem(item){
+            deleteItem(item, index, modal){
+                var isConfirm = confirm('Are you sure?');
+                if (!isConfirm) return false;
                 var vm = this
                 axios.delete(`/api/${this.resource}/${item}`)
                     .then(function (response) {
                         if (response.data.deleted) {
-
+                            modal.splice(index,1)
                         }
                     })
                     .catch(function (error) {
